Change fk naming scheme and check whether storage exists

+++ lib/dm-constraints/data_objects_adapter.rb
@@ -9,7 +9,7 @@ module DataMapper
             foreign_keys  = parent.key.map { |key| property_to_column_name(parent.repository(repository_name), key, false) } * ', '
             keys          = relationship.child_key.map { |key| property_to_column_name(model.repository(repository_name), key, false) } * ', '
             "ALTER TABLE #{quote_table_name(model.storage_name(repository_name))}
-             ADD CONSTRAINT #{foreign_table}_fk FOREIGN KEY (#{keys}) REFERENCES #{foreign_table} (#{foreign_keys})"
+             ADD CONSTRAINT #{relationship.name}_fk FOREIGN KEY (#{keys}) REFERENCES #{foreign_table} (#{foreign_keys})"
           end
         end
 
@@ -17,7 +17,7 @@ module DataMapper
           model.many_to_one_relationships.map do |relationship|
             foreign_table = relationship.parent_model.storage_name(repository_name)
             "ALTER TABLE #{quote_table_name(model.storage_name(repository_name))} 
-             DROP CONSTRAINT #{foreign_table}_fk"
+             DROP CONSTRAINT #{relationship.name}_fk"
           end
         end
 
@@ -36,9 +36,11 @@ module DataMapper
           def auto_migrate_constraints_down(repository_name, *descendants)
             descendants = DataMapper::Resource.descendants.to_a if descendants.empty?
             descendants.each do |model|
-              adapter = model.repository(repository_name).adapter
-              statements = adapter.destroy_constraints_statements(repository_name, model)
-              statements.each {|stmt| adapter.execute(stmt) }
+              if model.storage_exists?(repository_name)
+                adapter = model.repository(repository_name).adapter
+                statements = adapter.destroy_constraints_statements(repository_name, model)
+                statements.each {|stmt| adapter.execute(stmt) }
+              end
             end
           end
